// src/types/university.types.ts
export interface University {
  id: number;
  name: string;
  shortName: string;
  country: string;
  flag: string;
  ranking: number;
  image: string;
  tuition: string;
  acceptanceRate: string;
  scholarships: boolean;
  type: string;
  programs: string[];
  ieltsRequired: string;
  description: string;
}